+++ site/includes/class.inc.php
@@ -32,13 +32,13 @@
       * @since Pb2.0Bfx0RCA
       * @var string
       */
-      public $version = "Pb2.3Bfx0";
+      public $version = "Pb2.4Bfx0";
       /**
       * Version code
       * @since Pb2.0Bfx0RCA
       * @var int
       */
-      public $version_code = 20230900;
+      public $version_code = 20240900;
       /**
       * True when USOC is modded.
       * @since Pb2.0Bfx0RCA