+++ site/login/changepass.php
@@ -28,7 +28,7 @@
   if($passc){
     $sql = "UPDATE User SET password='".password_hash($_POST["newpass1"],PASSWORD_DEFAULT)."' WHERE Username='".$_SESSION['User_Name']."';";
     echo $sql;
-    $db_erg = mysqli_query( $U->db_link, $sql );
+    $db_erg = mysqli_query($U->db_link, $sql);
   }else{
     header("Location: ".$USOC["DOMAIN"]."/error");
   }